changes filters init from main to controller
